% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_functions_package-package.R
\docType{package}
\name{dk_functions_package-package}
\alias{dk_functions_package}
\alias{dk_functions_package-package}
\title{dk_functions_package: Personal Package for David Kingman}
\description{
A package containing David Kingman's personal utility functions and R templates.
}
\author{
\strong{Maintainer}: David Kingman \email{david_kingman@hotmail.co.uk}

}
\keyword{internal}
