% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_survey_mean.R
\name{robust_survey_mean}
\alias{robust_survey_mean}
\title{A tidyverse-friendly method of estimating robust population means and their uncertainty from survey datasets}
\usage{
robust_survey_mean(
  x,
  method = c("trim", "winsorize"),
  LB = 0,
  UB = 1,
  na.rm = FALSE,
  vartype = c("se", "ci", "var", "cv"),
  ...
)
}
\arguments{
\item{x}{A survey design object.}

\item{method}{Choose the method to use to estimate the robust mean from a choice of either 'trim' for a trimemd mean or 'winsorize' for a winsorised mean. Default is
'trim'.}

\item{LB}{The lower bound to either trim or winsorize. See the documentation for the original functions in robsurvey for more information.}

\item{UB}{The uper bound to either trim or winsorize. As above.}

\item{na.rm}{Remove observations which have missing data before estimating the mean? Default is FALSE.}

\item{vartype}{The estimate of uncertainty to return. Default is 'se' for standard error of the mean.}

\item{...}{Additional arguments to pass to robsurvey::svymean_trimmed or robsurvey::svymean_winsorized.}
}
\value{
A tibble containing the estimated population parameter.
}
\description{
This function is a wrapper around robsurvey::svymean_trimmed() and robsurvey::svymean_winsorized() which is designed to allow them to work harmoniously with the
syntax of the tidyverse, borrowing ideas from the srvyr package. It is designed to be called from within dplyr functions like dplyr::summarise() or dplyr::mutate().
}
